updating updateUser to second gen
