+++ functions/src/adminGetUserEmailSecondGen.js
@@ -6,19 +6,22 @@ const {
   InvalidRoleError,
 } = require("./lib/customErrors");
 
-exports.adminGetUserEmailSecondGen = onCall(async (request) => {
-  if (!request.auth) {
-    throw new UnauthenticatedError("The user is not authenticated.");
-  }
-  if (!request.auth.token.superadmin) {
-    throw new InvalidRoleError("Only Superadmin can see user email.");
-  }
-  try {
-    const userRecord = await getAuth().getUser(request.data.userID);
+exports.adminGetUserEmailSecondGen = onCall(
+  { region: "europe-north1" },
+  async (request) => {
+    if (!request.auth) {
+      throw new UnauthenticatedError("The user is not authenticated.");
+    }
+    if (!request.auth.token.superadmin) {
+      throw new InvalidRoleError("Only Superadmin can see user email.");
+    }
+    try {
+      const userRecord = await getAuth().getUser(request.data.userID);
 
-    return userRecord.email;
-  } catch (error) {
-    console.error("Error fetching user data:", error);
-    throw new HttpsError("internal", "Error fetching user data");
+      return userRecord.email;
+    } catch (error) {
+      console.error("Error fetching user data:", error);
+      throw new HttpsError("internal", "Error fetching user data");
+    }
   }
-});
+);
