+++ functions/src/updateUserSecondGen.js
@@ -1,51 +1,61 @@
-// const admin = require("firebase-admin");
-// const functions = require("firebase-functions");
+const { getAuth } = require("firebase-admin/auth");
+const { getFirestore } = require("firebase-admin/firestore");
+const { onCall, HttpsError } = require("firebase-functions/v2/https");
 
-// const {
-//   UnauthenticatedError,
-//   InvalidRoleError,
-//   NotPermittedError,
-// } = require("./lib/customErrors");
+const {
+  UnauthenticatedError,
+  InvalidRoleError,
+  NotPermittedError,
+} = require("./lib/customErrors");
 
-// exports.updateUserSecondGen = functions.https.onCall(
-//   async (changedUser, context) => {
-//     if (!context.auth) {
-//       throw new UnauthenticatedError("The user is not authenticated.");
-//     }
+exports.updateUserSecondGen = onCall(
+  { region: "europe-north1" },
+  async (request) => {
+    if (!request.auth) {
+      throw new UnauthenticatedError("The user is not authenticated.");
+    }
 
-//     if (!context.auth.token.superadmin) {
-//       throw new InvalidRoleError("Only Superadmin can change user claims");
-//     }
+    const token = request.auth.token;
 
-//     if (changedUser.id === changedUser.adminID) {
-//       throw new NotPermittedError(
-//         "Superadmin cannot become admin of himself/herself."
-//       );
-//     }
-//     try {
-//       const claims = {};
-//       claims[changedUser.role] = true;
+    if (!token.superadmin) {
+      throw new InvalidRoleError("Only Superadmin can change user claims");
+    }
 
-//       await admin
-//         .auth()
-//         .setCustomUserClaims(changedUser.id, claims)
-//         .then(() => {
-//           admin.firestore().collection("Users").doc(changedUser.id).update({
-//             first_name: changedUser.firstName,
-//             last_name: changedUser.lastName,
-//             status_active: changedUser.statusActive,
-//             role: changedUser.role,
-//             admin_id: changedUser.adminID,
-//           });
-//         });
+    const changedUser = request.data;
 
-//       return { success: true };
-//     } catch (error) {
-//       console.error("Error fetching user data:", error);
-//       throw new functions.https.HttpsError(
-//         "internal",
-//         "Error fetching user data"
-//       );
-//     }
-//   }
-// );
+    if (changedUser.id === changedUser.adminID) {
+      throw new NotPermittedError(
+        "Superadmin cannot become admin of himself/herself."
+      );
+    }
+    try {
+      const claims = {};
+      claims[changedUser.role] = true;
+
+      await getAuth()
+        .setCustomUserClaims(changedUser.id, claims)
+        .then(() => {
+          getFirestore().collection("Users").doc(changedUser.id).update({
+            first_name: changedUser.firstName,
+            last_name: changedUser.lastName,
+            status_active: changedUser.statusActive,
+            role: changedUser.role,
+            admin_id: changedUser.adminID,
+          });
+        });
+
+      return { success: true };
+    } catch (error) {
+      if (error.type === "UnauthenticatedError") {
+        throw new HttpsError("unauthenticated", error.message);
+      } else if (
+        error.type === "NotAnAdminError" ||
+        error.type === "InvalidRoleError"
+      ) {
+        throw new HttpsError("failed-precondition", error.message);
+      } else {
+        throw new HttpsError("internal", error.message);
+      }
+    }
+  }
+);
