+++ src/firebase-functions/updateTS/superAdminUpdatesUserInfo.ts
@@ -3,7 +3,9 @@ import functions from "@react-native-firebase/functions";
 
 export const superAdminUpdatesUserInfo = async (user: User) => {
   try {
-    const updateUser = functions().httpsCallable("updateUser");
+    const updateUser = functions().httpsCallableFromUrl(
+      "https://europe-north1-dev-do-good-get-good.cloudfunctions.net/updateUserSecondGen",
+    );
     const res = await updateUser(user);
     console.log("Super admin update user data ");
     return res.data;
